requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking requirements.in
#
asgiref==3.6.0
    # via
    #   django
    #   django-countries
attrs==22.2.0
    # via pytest
autoflake==1.4
    # via -r requirements.in
black==22.3.0
    # via -r requirements.in
bleach[css]==5.0.1
    # via django-bleach
boto3==1.26.85
    # via django-storages
botocore==1.29.85
    # via
    #   boto3
    #   s3transfer
build==0.10.0
    # via pip-tools
cachetools==5.3.0
    # via google-auth
certifi==2022.12.7
    # via
    #   requests
    #   sentry-sdk
cfgv==3.3.1
    # via pre-commit
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   black
    #   django-click
    #   pip-tools
coverage==7.2.1
    # via pytest-cov
cssselect==1.2.0
    # via pyquery
distlib==0.3.6
    # via virtualenv
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2.19
    # via
    #   -r requirements.in
    #   django-admin-sortable2
    #   django-bleach
    #   django-debug-toolbar
    #   django-extensions
    #   django-formtools
    #   django-leaflet
    #   django-recaptcha
    #   django-sendgrid-v5
    #   django-storages
    #   django-unused-media
    #   easy-thumbnails
django-admin-sortable2==1.0.4
    # via -r requirements.in
django-autoslug==1.9.4
    # via -r requirements.in
django-bleach==3.0.1
    # via -r requirements.in
django-click==2.3.0
    # via -r requirements.in
django-countries==7.5.1
    # via -r requirements.in
django-date-extensions==3.1.2
    # via -r requirements.in
django-debug-toolbar==3.8.1
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-formtools==2.3
    # via -r requirements.in
django-gulp-rev==0.2
    # via -r requirements.in
django-leaflet==0.24.0
    # via -r requirements.in
django-markdown-deux==1.0.5
    # via -r requirements.in
django-recaptcha==3.0.0
    # via -r requirements.in
django-sendgrid-v5==0.8.1
    # via -r requirements.in
django-storages[boto3]==1.7.1
    # via -r requirements.in
django-tinymce==3.5.0
    # via -r requirements.in
django-unused-media==0.1.13
    # via -r requirements.in
dnspython==2.3.0
    # via -r requirements.in
easy-thumbnails==2.8.5
    # via -r requirements.in
easydict==1.10
    # via -r requirements.in
filelock==3.9.0
    # via
    #   urlextract
    #   virtualenv
flake8==4.0.1
    # via -r requirements.in
freezegun==1.2.2
    # via -r requirements.in
future==0.18.3
    # via django-sendgrid-v5
google-api-core==2.11.0
    # via google-api-python-client
google-api-python-client==2.37.0
    # via -r requirements.in
google-auth==2.16.2
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.1.0
    # via google-api-python-client
googleapis-common-protos==1.58.0
    # via google-api-core
httplib2==0.21.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
icalendar==4.0.3
    # via -r requirements.in
identify==2.5.18
    # via pre-commit
idna==3.4
    # via
    #   requests
    #   urlextract
    #   yarl
iniconfig==2.0.0
    # via pytest
isort==5.8.0
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lxml==4.9.2
    # via pyquery
markdown2==2.4.8
    # via django-markdown-deux
mccabe==0.6.1
    # via flake8
multidict==6.0.4
    # via yarl
mypy-extensions==1.0.0
    # via black
nodeenv==1.7.0
    # via pre-commit
oauth2client==4.1.3
    # via -r requirements.in
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.0
    # via
    #   build
    #   pytest
pathspec==0.11.0
    # via black
pillow==9.3.0
    # via
    #   -r requirements.in
    #   easy-thumbnails
pip-tools==6.12.1
    # via -r requirements.in
platformdirs==3.1.0
    # via
    #   black
    #   urlextract
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.21.0
    # via -r requirements.in
protobuf==4.22.0
    # via
    #   google-api-core
    #   googleapis-common-protos
psycopg2-binary==2.8.3
    # via -r requirements.in
py==1.11.0
    # via pytest
py-trello==0.13.0
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via
    #   google-auth
    #   oauth2client
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via
    #   autoflake
    #   flake8
pyparsing==3.0.9
    # via httplib2
pyproject-hooks==1.0.0
    # via build
pyquery==1.4.0
    # via -r requirements.in
pytest==6.2.5
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-dotenv
    #   pytest-env
    #   pytest-mock
pytest-cov==2.12.1
    # via -r requirements.in
pytest-django==4.4.0
    # via -r requirements.in
pytest-dotenv==0.5.2
    # via -r requirements.in
pytest-env==0.6.2
    # via -r requirements.in
pytest-mock==3.10.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   freezegun
    #   icalendar
    #   py-trello
python-dotenv==0.21.0
    # via
    #   -r requirements.in
    #   pytest-dotenv
python-http-client==3.3.7
    # via
    #   django-sendgrid-v5
    #   sendgrid
pytz==2022.7.1
    # via
    #   django
    #   icalendar
    #   py-trello
pyyaml==6.0
    # via
    #   pre-commit
    #   vcrpy
requests==2.28.2
    # via
    #   google-api-core
    #   py-trello
    #   requests-oauthlib
    #   stripe
requests-oauthlib==1.3.1
    # via py-trello
rsa==4.9
    # via
    #   google-auth
    #   oauth2client
s3transfer==0.6.0
    # via boto3
sendgrid==6.9.7
    # via django-sendgrid-v5
sentry-sdk==1.16.0
    # via -r requirements.in
six==1.16.0
    # via
    #   bleach
    #   django-unused-media
    #   google-auth
    #   google-auth-httplib2
    #   oauth2client
    #   python-dateutil
    #   vcrpy
slack-sdk==3.19.5
    # via -r requirements.in
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
starkbank-ecdsa==2.2.0
    # via sendgrid
stripe==5.2.0
    # via -r requirements.in
tinycss2==1.1.1
    # via bleach
toml==0.10.2
    # via
    #   pytest
    #   pytest-cov
typing-extensions==4.5.0
    # via django-countries
uritemplate==4.1.1
    # via google-api-python-client
uritools==4.0.1
    # via urlextract
urlextract==1.8.0
    # via -r requirements.in
urllib3==1.26.14
    # via
    #   botocore
    #   requests
    #   sentry-sdk
vcrpy==4.2.1
    # via -r requirements.in
virtualenv==20.20.0
    # via pre-commit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
wheel==0.38.4
    # via pip-tools
wrapt==1.15.0
    # via vcrpy
yarl==1.8.2
    # via vcrpy

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
